docs: note ui modules
